+++ CHANGELOG.md
@@ -7,6 +7,16 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+## [9.86.3] - 2019-10-07
+
+### Added
+
+- Checkboxes to `EXPERIMENTAL_TableTree`
+
+### Changed
+
+- Separate `EXPERIMENTAL_Table`'s BulkContext from TableContext
+
 ## [9.86.2] - 2019-10-04
 
 ### Fixed